Adds indirect energy to the EROI values

To add indirect energy, the function calculates the ratio of indirect energy to output per product group at the global level (see details), and then adds it to the energy inputs using the inverse of the EROI. The function uses the same ratio for each product group, independently of whether the energy stage is electricity, heat, fuel, or the combination (it does apply a different ratio at the primary, final and useful stage though).

Usage

add_indirect_energy_to_erois(
  .tidy_summarised_erois_df,
  .tidy_indirect_energy,
  .tidy_iea_df,
  country = IEATools::iea_cols$country,
  year = IEATools::iea_cols$year,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  e_dot = IEATools::iea_cols$e_dot,
  unit = IEATools::iea_cols$unit,
  product.group = "Product.Group",
  energy.stage = "Energy.stage",
  group.eroi = "Group.eroi",
  total_group_output = "Total_Group_Output",
  indirect_energy_ktoe = "Indirect_Energy_ktoe",
  eroi.method = "Eroi.method",
  type = "Type",
  boundary = "Boundary",
  final_to_useful_eff = "Final_to_useful_efficiency",
  ratio_indirect_energy_per_output = "ratio_indirect_energy_per_output",
  include_non_energy_uses = TRUE
)

Arguments

.tidy_summarised_erois_df

The aggregated EROI values to which indirect energy needs to be added

.tidy_indirect_energy

A tidy data frame containing the indirect energy to be added, in ktoe.

.tidy_iea_df

The .tidy_iea_df from which the output per product group is calculated.

country, year, method, energy_type, last_stage, e_dot, unit

See IEATools::iea_cols.

product.group

The name of the column containing the product group name. Default is "Product.Group".

energy.stage

The name of the column containing the energy stage for the calculation of the EROI. Default is "Energy.stage".

group.eroi

The name of the column containing the EROI values. Default is "Group.eroi".

total_group_output

The name of the column containing the total product group output. Default is "Total_Group_Output".

indirect_energy_ktoe

The name of the column containing the indirect energy data in the input indirect energy data frame. Default is "Indirect_Energy_ktoe".

eroi.method

The name of the column containing the eroi method. Default is "Eroi.method".

type

The name of the column containing the type of eroi calculated. Default is "Type".

boundary

The name of the column containing the boundary for the eroi calculation. Default is "Boundary".

final_to_useful_eff

The name of the temporary column containing the calculated final-to-useful efficiencies. Default is "Final_to_useful_efficiency".

ratio_indirect_energy_per_output

The name of the column containing the ratio of indirect energy to product group output. Default is "ratio_indirect_energy_per_output".

include_non_energy_uses

A boolean stating whether non-energy uses are included in the calculation of the output per product group.

Value

A data frame containing the EROIs with the indirect energy now included.

Details

The .tidy_iea_data frame provided as input should be a representation of the global Energy Conversion Chain, as the ratio of indirect energy to output (primary and final energy stages) is calculated at the global level, so output per fossil fuel group needs to be calculated at the global level.