Calculates the share of heat provided by each fossil fuel group

This function calculates, for each country, the share of heat that has been supplied by each fossil fuel group, out of the domestically produced heat. So, shares don't have to add up to unity - indeed they won't if any heat is supplied by renewables or nuclear.

Usage

calc_share_heat_supply_by_ff_group(
  .tidy_iea_df,
  list_supply_mats = c(IEATools::psut_cols$V),
  list_oil_products = IEATools::oil_and_oil_products,
  list_coal_products = IEATools::coal_and_coal_products,
  list_gas_products = IEATools::primary_gas_products,
  exports = IEATools::interface_industries$exports,
  losses = IEATools::tfc_compare_flows$losses,
  country = IEATools::iea_cols$country,
  method = IEATools::iea_cols$method,
  energy_type = IEATools::iea_cols$energy_type,
  last_stage = IEATools::iea_cols$last_stage,
  year = IEATools::iea_cols$year,
  product = IEATools::iea_cols$product,
  unit = IEATools::iea_cols$unit,
  flow = IEATools::iea_cols$flow,
  e_dot = IEATools::iea_cols$e_dot,
  matnames = IEATools::mat_meta_cols$matnames,
  product.group = "Product.Group",
  energy.stage = "Energy.stage",
  share = "Share",
  product_without_origin = "product_without_origin"
)

Arguments

.tidy_iea_df

The .tidy_iea_df for which the share of heat provided by each fossil fuel group needs to be calculated.

list_supply_mats

The list of the supply matrices to be used for the calculating the shares. Default is c(IEATools::psut_cols$V).

list_oil_products

The list of oil products to be used when calculating the use shares. Default is IEATools::oil_and_oil_products.

list_coal_products

The list of coal products to be used when calculating the use shares. Default is IEATools::coal_and_coal_products.

list_gas_products

The list of gas products to be used when calculating the use shares. Default is IEATools::primary_gas_products.

exports

The character string identifying exports flows in the .tidy_iea_df. Default is IEATools::interface_industries$exports.

losses

The character string identifying losses flows in the .tidy_iea_df. Default is IEATools::tfc_compare_flows$losses.

country, method, energy_type, last_stage, year, product, unit, flow, e_dot

See IEATools::iea_cols.

matnames

The name of the column containing the matrices names. Default is IEATools::mat_meta_cols$matnames.

product.group

The name of the column containing the product group name. Default is "Product.Group".

energy.stage

The name of the column containing the energy stage. Default is "Energy.stage".

share

The name of the column containing the shares of heat delivered by each fossil fuel group.

product_without_origin

The name of the column containing the product name without the product origin. Default is "product_without_origin"

Value

A tidy data frame containing the shares of heat delivered by each fossil fuel group.

Details

Fossil fuel groups used are "All fossil fuels", "Oil and gas products", "Oil products", "Natural gas", and "Coal products", so they are obviously not exclusive. The function can work both on a single country Energy Conversion Chain of Domestic Technology Assumption type, or with a multi-regional Energy Conversion Chain for instance using the Global Market Assumption. The input data frame will have to be slightly adapted in this case (for an example see the tests related to the function)

Examples

ECCTools::tidy_AB_data %>% 
   IEATools::add_psut_matnames() %>% 
   ECCTools::transform_to_dta() %>% 
   calc_share_heat_supply_by_ff_group()
#> # A tibble: 0 × 9
#> # Groups:   Country, Method, Energy.type, Last.stage, Year, Unit [0]
#> # ℹ 9 variables: Country <chr>, Method <chr>, Energy.type <chr>,
#> #   Last.stage <chr>, Year <dbl>, Unit <chr>, Product.Group <chr>,
#> #   Energy.stage <chr>, Share <dbl>